Add files via upload
These are all of the base files of how the project was left. I was inspired by Thingiverse and if I were to revisit this project sometime in the future I would not want to write it completely from scratch. Instead I'd probably use a combination of React, Bootstrap and Kubernetes to make a better product.